refactoring & backwards compatability
