+++ src/ibex_bluesky_core/devices/dae/strategies/__init__.py
@@ -18,10 +18,7 @@ from ibex_bluesky_core.devices.dae.strategies._reducers import (
     MonitorNormalizer,
     PeriodGoodFramesNormalizer,
     PeriodSpecIntegralsReducer,
-    PolarisingReducer,
     ScalarNormalizer,
-    WavelengthBoundedNormalizer,
-    polarization,
     sum_spectra,
     tof_bounded_spectra,
     wavelength_bounded_spectra,
@@ -48,7 +45,6 @@ __all__ = [
     "PeriodGoodFramesWaiter",
     "PeriodPerPointController",
     "PeriodSpecIntegralsReducer",
-    "PolarisingReducer",
     "ProvidesExtraReadables",
     "Reducer",
     "RunPerPointController",
@@ -56,9 +52,7 @@ __all__ = [
     "SimpleWaiter",
     "TimeWaiter",
     "Waiter",
-    "WavelengthBoundedNormalizer",
     "_end_or_abort_run",
-    "polarization",
     "sum_spectra",
     "tof_bounded_spectra",
     "wavelength_bounded_spectra",
